Put everything into one message
Turns out that clients other than transmission tend to not work very
well in multithreaded environment.  Since I want to be compatible with
as much torrents clients as possible I am dropping support altogether

As a consequence call_on_finish is no longer support.  It should be
handled by torrent client anyway.  This was beyond the scope of this
project.

In addition, now all torrents are tracked every 3 seconds and code is
much simpler.